dashboard: convert the dashboard page to a server page.
