+++ app/(main)/admin/dashboard/components/adminDashboard.tsx
@@ -1,0 +1,356 @@
+"use client" 
+import {Tabs, TabsContent} from "@/components/ui/tabs"
+import {
+    Card,
+    CardContent,
+} from "@/components/ui/card"
+import useDashboard from "../hooks/useDahboard"
+import { IDashboardForm, IFormCreationData, IFormResponseData, IUserData, IUsersActivityData } from "@/@types"
+import UsersTable from "./tables/usersTable"
+import FormsTable from "./tables/formsTable"
+import SearchBar from "../components/searchBar"
+import AllTabs from "../components/allTabs"
+import { TABS } from "../constants/constants"
+import TabHeader from "../components/tabHeader"
+import AllCharts from "../components/allCharts"
+import { useMemo } from "react"
+import AllCards from "../components/allStates"
+import DashboardHeader from "../components/dashboardHeader"
+interface IProps {
+    totalUsers: number ,
+    activeUsers: number,
+    totalForms: number,
+    totalResponses: number,
+    usersData: IUserData[],
+    formsData: IDashboardForm[],
+    userActivityData: IUsersActivityData[],
+    formCreationData: IFormCreationData[],
+    formResponsesData: IFormResponseData[]
+}
+// Mock data for the dashboard
+export const mockUserData: IUserData[] = [
+    {
+            id: 1,
+            name: "John Doe",
+            email: "john@example.com",
+            status: "active",
+            role: "admin",
+            forms: 12,
+            lastActive: "2025-04-18",
+        },
+        {
+            id: 2,
+            name: "Jane Smith",
+            email: "jane@example.com",
+            status: "active",
+            role: "creator",
+            forms: 8,
+            lastActive: "2025-04-17",
+        },
+        {
+            id: 3,
+            name: "Robert Johnson",
+            email: "robert@example.com",
+            status: "inactive",
+            role: "user",
+            forms: 3,
+            lastActive: "2025-04-10",
+        },
+        {
+            id: 4,
+            name: "Emily Davis",
+            email: "emily@example.com",
+            status: "active",
+            role: "creator",
+            forms: 15,
+            lastActive: "2025-04-18",
+        },
+        {
+            id: 5,
+            name: "Michael Wilson",
+            email: "michael@example.com",
+            status: "active",
+            role: "user",
+            forms: 6,
+            lastActive: "2025-04-16",
+        },
+        {
+            id: 6,
+            name: "Sarah Brown",
+            email: "sarah@example.com",
+            status: "inactive",
+            role: "user",
+            forms: 0,
+            lastActive: "2025-04-05",
+        },
+        {
+            id: 7,
+            name: "David Miller",
+            email: "david@example.com",
+            status: "active",
+            role: "creator",
+            forms: 9,
+            lastActive: "2025-04-17",
+        },
+        {
+            id: 8,
+            name: "Lisa Garcia",
+            email: "lisa@example.com",
+            status: "active",
+            role: "admin",
+            forms: 4,
+            lastActive: "2025-04-15",
+        },
+];
+
+const mockFormData = [
+    {
+        id: 1,
+        name: "Contact Form",
+        creator: "John Doe",
+        responses: 145,
+        createdAt: "2025-04-10"
+    },
+    {
+        id: 2,
+        name: "Job Application",
+        creator: "Emily Davis",
+        responses: 78,
+        createdAt: "2025-04-12"
+    },
+    {
+        id: 3,
+        name: "Customer Survey",
+        creator: "Jane Smith",
+        responses: 213,
+        createdAt: "2025-04-08"
+    },
+    {
+        id: 4,
+        name: "Event Registration",
+        creator: "Michael Wilson",
+        responses: 56,
+        createdAt: "2025-04-15"
+    }, {
+        id: 5,
+        name: "Feedback Form",
+        creator: "David Miller",
+        responses: 92,
+        createdAt: "2025-04-11"
+    }, {
+        id: 6,
+        name: "Newsletter Signup",
+        creator: "John Doe",
+        responses: 187,
+        createdAt: "2025-04-09"
+    }, {
+        id: 7,
+        name: "Product Order",
+        creator: "Emily Davis",
+        responses: 124,
+        createdAt: "2025-04-14"
+    }, {
+        id: 8,
+        name: "Support Request",
+        creator: "Lisa Garcia",
+        responses: 67,
+        createdAt: "2025-04-16"
+    },
+]
+
+// Chart data
+const userActivityData = [
+    {
+        name: "Mon",
+        active: 120,
+        new: 15
+    },
+    {
+        name: "Tue",
+        active: 132,
+        new: 12
+    },
+    {
+        name: "Wed",
+        active: 141,
+        new: 18
+    },
+    {
+        name: "Thu",
+        active: 158,
+        new: 24
+    }, {
+        name: "Fri",
+        active: 142,
+        new: 16
+    }, {
+        name: "Sat",
+        active: 90,
+        new: 8
+    }, {
+        name: "Sun",
+        active: 85,
+        new: 7
+    },
+]
+
+const formCreationData = [
+    {
+        name: "Jan",
+        forms: 65
+    },
+    {
+        name: "Feb",
+        forms: 78
+    },
+    {
+        name: "Mar",
+        forms: 90
+    },
+    {
+        name: "Apr",
+        forms: 112
+    }, {
+        name: "May",
+        forms: 130
+    }, {
+        name: "Jun",
+        forms: 142
+    }, {
+        name: "Jul",
+        forms: 151
+    }, {
+        name: "Aug",
+        forms: 160
+    }, {
+        name: "Sep",
+        forms: 175
+    }, {
+        name: "Oct",
+        forms: 190
+    }, {
+        name: "Nov",
+        forms: 205
+    }, {
+        name: "Dec",
+        forms: 220
+    },
+]
+
+const formResponsesData = [
+    {
+        name: "Contact Form",
+        value: 145
+    },
+    {
+        name: "Job Application",
+        value: 78
+    },
+    {
+        name: "Customer Survey",
+        value: 213
+    },
+    {
+        name: "Event Registration",
+        value: 56
+    }, {
+        name: "Feedback Form",
+        value: 92
+    },
+]
+
+
+const AdminDashboard = () => {
+    const {
+        filteredUsers,
+        filteredForms,
+        totalUsers,
+        searchUsers,
+        searchForms,
+        activeUsers,
+        totalForms,
+        totalResponses,
+        setSearchUsers,
+        setSearchForms,
+    } = useDashboard({mockUserData, mockFormData});
+
+    const stateCardsData = useMemo(() => [
+        {
+            stateTitle: "Total Users",
+            stateValue: totalUsers,
+            statePercentage: 10,
+        },
+        {
+            stateTitle: "Active Users",
+            stateValue: activeUsers,
+            statePercentage: 10,
+        },
+        {
+            stateTitle: "Total Forms",
+            stateValue: totalForms,
+            statePercentage: 15,
+        },
+        {
+            stateTitle: "Form Responses",
+            stateValue: totalResponses,
+            statePercentage: 3,
+        },
+    ], [totalUsers, activeUsers, totalForms, totalResponses]);
+    return (
+        <div className="min-h-screen bg-gradient-to-br from-purple-50 to-purple-100 pb-20 pt-0 md:pt-0 w-full">
+            <DashboardHeader/>
+
+            <div className="container mx-auto px-4 py-8">
+                {/* Dashboard Tabs */}
+                <Tabs defaultValue="overview" className="space-y-6">
+
+                    <AllTabs tabs={TABS}/>
+
+                    <TabsContent value="overview" className="space-y-6">
+
+                        <AllCards cards={stateCardsData} />
+
+                        <AllCharts
+                            formResponsesData={formResponsesData} 
+                            formCreationData={formCreationData} 
+                            userActivityData={userActivityData}
+                        /> 
+
+                    </TabsContent>
+
+                    <TabsContent value="users" className="space-y-6">
+                        <Card>
+                            <TabHeader title={"User Management"} description={"View and manage all users on the platform"}/>
+                            <CardContent>
+
+                                <SearchBar
+                                    placeholder="Search users..."
+                                    setSearch={setSearchUsers}
+                                    search={searchUsers}
+                                />
+
+                                <UsersTable filteredUsers={filteredUsers}/>
+                            </CardContent>
+                        </Card>
+                    </TabsContent>
+
+                    <TabsContent value="forms" className="space-y-6">
+                        <Card>
+                            <TabHeader title={"Form Analytics"} description={"View and analyze all forms on the platform"}/>
+                            <CardContent>
+                                <SearchBar
+                                    placeholder="Search forms..."
+                                    setSearch={setSearchForms}
+                                    search={searchForms}
+                                />
+                                <FormsTable filteredForms={filteredForms} />                            
+                            </CardContent>
+                        </Card>
+                    </TabsContent>
+                </Tabs>
+            </div>
+        </div>
+    )
+}
+
+export default AdminDashboard;
