+++ app/(main)/admin/dashboard/components/charts/latestUsers.tsx
@@ -9,7 +9,9 @@ import {
 import {Badge} from "@/components/ui/badge"
 import {Clock, FileText} from "lucide-react"
 
-// Mock data for active users - replace with your actual data source
+interface Iprops {
+    
+}
 const activeUsers = [
     {
         id: 1,
