+++ app/(main)/admin/dashboard/components/dashboardHeader.tsx
@@ -1,0 +1,13 @@
+import React from 'react'
+
+const DashboardHeader = () => {
+    return (
+            <header className="bg-white border-b border-purple-100 shadow-sm sticky top-0 z-10 mt-14 md:mt-0">
+                <div className="container mx-auto px-4 py-4">
+                    <h1 className="text-xl font-bold text-purple-900">Admin Dashboard</h1>
+                </div>
+            </header>
+    )
+}
+
+export default DashboardHeader
